Add files via upload
The System has a tree visualization on web browsers and a back-end MySql 3NF Database.  A Python REST service fetches and stores data. The front-end is javascript and d3.js library for interactive visualizations. The Tree has nodes, sub-nodes and links between them. Each node represents a subject or a subsidiary subject. The link between two nodes represents a relationship. Any two subjects can have a relationship like Anga (part-whole), Aadhaara (supported-supporting), and others. These relationships are specified according to Indian categories described in the Shastras. Thus it is an Ontological schematic according to ancient Indian systems.  

The knowledge of an object is said to involve three ingredients, known in Sanskrit as Pramatr, Pramāņa and Praméya. The word Pramatr means the perceiver, the cogniser, or the knower. Pramāņa is the process of knowing. Praméya is the end-result of the knowledge process - i.e., the object that is known. There is something or someone that k